+++ downloads/remote_registry.go
@@ -1,5 +1,5 @@
 // DBDeployer - The MySQL Sandbox
-// Copyright © 2006-2019 Giuseppe Maxia
+// Copyright © 2006-2020 Giuseppe Maxia
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -373,6 +373,7 @@ func GetTarballInfo(fileName string, description TarballDescription) (TarballDes
 
 func checkRemoteUrl(remoteUrl string) error {
 
+	// #nosec G107
 	resp, err := http.Get(remoteUrl)
 	if err != nil {
 		return fmt.Errorf("[checkRemoteUrl] error getting %s: %s", remoteUrl, err)