+++ test/all_tests.sh
@@ -118,6 +118,7 @@ function all_tests {
         run_test ./test/go-unit-tests.sh
         run_test ./test/functional-test.sh
         run_test ./test/docker-test.sh $version
+        run_test ./test/cookbook-test.sh
     fi
     run_test ./test/mock/defaults-change.sh
     run_test ./test/mock/short-versions.sh